Move Stucts to main file
