+++ malbeacon/src/requirements.txt
@@ -1,3 +1,5 @@
 git+https://github.com/OpenCTI-Platform/client-python@master
 PyYAML==5.3.1
 requests==2.24.0
+pydantic==1.6.1
+urllib3==1.25.8
